Converted user wallet holdings logic into dedicated hook
